Move expr_skeleton class to its own file
Makes it easier to navigate through the code.
